Modify code to go back to index.html
Apparently using <button> to allow users to go back to index.html is not
a good practice, as users wouldn't be able to right click and to be able
to open it in new tab.

Therefore, I implemented the use of anchor tag instead (<a></a>) and
style it in CSS in the near future.